requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.1.1
django-phonenumber-field==5.0.0
django-user-agents==0.4.0
filelock==3.0.12
gunicorn==20.0.4
phonenumbers==8.12.9
Pillow==7.2.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
ua-parser==0.10.0
user-agents==2.2.0
virtualenv==20.0.31
